% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elisa_calibration}
\alias{elisa_calibration}
\title{Package example Elisa data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.}
\usage{
elisa_calibration
}
\description{
Package example Elisa data set
}
\keyword{datasets}
